Conflicts:
	src/main/java/com/cloudstone/emenu/logic/OrderLogic.java

+++ src/main/java/com/cloudstone/emenu/logic/MenuLogic.java
@@ -4,19 +4,26 @@
  */
 package com.cloudstone.emenu.logic;
 
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import com.cloudstone.emenu.data.Chapter;
 import com.cloudstone.emenu.data.Dish;
+import com.cloudstone.emenu.data.DishNote;
 import com.cloudstone.emenu.data.DishTag;
 import com.cloudstone.emenu.data.IdName;
 import com.cloudstone.emenu.data.Menu;
 import com.cloudstone.emenu.data.MenuPage;
+import com.cloudstone.emenu.exception.BadRequestError;
 import com.cloudstone.emenu.exception.DataConflictException;
-import com.cloudstone.emenu.service.MenuService;
+import com.cloudstone.emenu.service.IMenuService;
+import com.cloudstone.emenu.storage.db.IDishPageDb.DishPage;
+import com.cloudstone.emenu.util.CollectionUtils;
+import com.cloudstone.emenu.util.CollectionUtils.Tester;
 import com.cloudstone.emenu.util.DataUtils;
 import com.cloudstone.emenu.util.StringUtils;
 
@@ -27,7 +34,7 @@ import com.cloudstone.emenu.util.StringUtils;
 @Component
 public class MenuLogic extends BaseLogic {
     @Autowired
-    private MenuService menuService;
+    private IMenuService menuService;
     
     @Autowired
     private ImageLogic imageLogic;
@@ -68,6 +75,15 @@ public class MenuLogic extends BaseLogic {
         List<Menu> menus = menuService.getAllMenu();
         DataUtils.filterDeleted(menus);
         return menus;
+    }
+    
+    public Menu getCurrentMenu() {
+        //TODO
+        List<Menu> menus = getAllMenu();
+        if (menus.size() > 0) {
+            return menus.get(0);
+        }
+        return null;
     }
     
     public Menu updateMenu(Menu menu) {
@@ -182,6 +198,29 @@ public class MenuLogic extends BaseLogic {
         return menuService.getChapter(id);
     }
     
+    public List<Chapter> listChapters(final int menuId, int dishId) {
+        List<MenuPage> pages = listMenuPageByDishId(dishId);
+        Set<Integer> chapterIds = new HashSet<Integer>();
+        for (MenuPage p:pages) {
+            chapterIds.add(p.getChapterId());
+        }
+        List<Chapter> chapters = menuService.listChapter(CollectionUtils.toIntArray(chapterIds));
+        DataUtils.filterDeleted(chapters);
+        CollectionUtils.filter(chapters, new Tester<Chapter>() {
+            @Override
+            public boolean test(Chapter data) {
+                return data.getMenuId() == menuId;
+            }
+        });
+        return chapters;
+    }
+    
+    public List<DishPage> listDishPage(int dishId) {
+        List<DishPage> dishPages = menuService.listDishPage(dishId);
+        DataUtils.filterDeleted(dishPages);
+        return dishPages;
+    }
+    
     public List<Chapter> getAllChapter() {
         List<Chapter> chapters = menuService.getAllChapter();
         DataUtils.filterDeleted(chapters);
@@ -209,14 +248,45 @@ public class MenuLogic extends BaseLogic {
         return chapters;
     }
     
+    /* ---------- MenuPage ---------- */
+    
+    public List<MenuPage> listMenuPageByDishId(int dishId) {
+        List<DishPage> relations = listDishPage(dishId);
+        Set<Integer> pageIds = new HashSet<Integer>();
+        for (DishPage r:relations) {
+            pageIds.add(r.getMenuPageId());
+        }
+        List<MenuPage> pages = menuService.listMenuPage(CollectionUtils.toIntArray(pageIds));
+        DataUtils.filterDeleted(pages);
+        return pages;
+    }
+    
     public List<MenuPage> listMenuPageByChapterId(int chapterId) {
         List<MenuPage> datas = menuService.listMenuPageByChapterId(chapterId);
         DataUtils.filterDeleted(datas);
+        
+        for (int i=1; i<=datas.size(); i++) {
+            MenuPage p = datas.get(i-1);
+            if (p.getOrdinal() != i) {
+                p.setOrdinal(i);
+                innnerUpdateMenuPage(p);
+            }
+        }
         return datas;
     }
     
-    /* ---------- MenuPage ---------- */
+    private void innnerUpdateMenuPage(MenuPage p) {
+        p.setUpdateTime(System.currentTimeMillis());
+        menuService.updateMenuPage(p);
+    }
+    
     public MenuPage addMenuPage(MenuPage page) {
+        List<MenuPage> pages = listMenuPageByChapterId(page.getChapterId());
+        for (int i=page.getOrdinal(); i<=pages.size(); i++) {
+            MenuPage p = pages.get(i-1);
+            p.setOrdinal(i+1);
+            innnerUpdateMenuPage(p);
+        }
         long now = System.currentTimeMillis();
         page.setCreatedTime(now);
         page.setUpdateTime(now);
@@ -225,17 +295,50 @@ public class MenuLogic extends BaseLogic {
     }
     
     public void deleteMenuPage(int id) {
+        MenuPage old = getMenuPage(id);
+        if (old==null || old.isDeleted()) {
+            throw new BadRequestError();
+        }
+        List<MenuPage> pages = listMenuPageByChapterId(old.getChapterId());
+        for (int i=old.getOrdinal()+1; i<=pages.size(); i++) {
+            MenuPage p = pages.get(i-1);
+            p.setOrdinal(p.getOrdinal()-1);
+            innnerUpdateMenuPage(p);
+        }
         menuService.deleteMenuPage(id);
     }
     
-    public List<MenuPage> listMenuPage(int chapterId) {
-        List<MenuPage> pages = menuService.listMenuPageByChapterId(chapterId);
-        DataUtils.filterDeleted(pages);
-        return pages;
+    public MenuPage getMenuPage(int id) {
+        return menuService.getMenuPage(id);
+    }
+    
+    public MenuPage updateMenuPage(MenuPage page) {
+        MenuPage old = getMenuPage(page.getId());
+        if (old == null) {
+            throw new BadRequestError();
+        }
+        if (old.getOrdinal() != page.getOrdinal()) {
+            List<MenuPage> pages = listMenuPageByChapterId(page.getChapterId());
+            if (page.getOrdinal() < old.getOrdinal()) {
+                for (int i=page.getOrdinal(); i<=old.getOrdinal()-1; i++) {
+                    MenuPage p = pages.get(i-1);
+                    p.setOrdinal(p.getOrdinal()+1);
+                    innnerUpdateMenuPage(p);
+                }
+            } else {
+                for (int i=old.getOrdinal()+1; i<=page.getOrdinal(); i++) {
+                    MenuPage p = pages.get(i-1);
+                    p.setOrdinal(p.getOrdinal()-1);
+                    innnerUpdateMenuPage(p);
+                }
+            }
+        }
+        innnerUpdateMenuPage(page);
+        return menuService.getMenuPage(page.getId());
     }
     
     /* ---------- DishTag ---------- */
-    public List<DishTag> listAllDisTag() {
+    public List<DishTag> listAllDishTag() {
         List<DishTag> tags = menuService.listAllDishTag();
         DataUtils.filterDeleted(tags);
         return tags;
@@ -269,5 +372,42 @@ public class MenuLogic extends BaseLogic {
     }
     public void deleteDishTag(int id) {
         menuService.deleteDishTag(id);
+    }
+    
+    /* ---------- DishNote ---------- */
+    public List<DishNote> listAllDishNote() {
+        List<DishNote> notes = menuService.listAllDishNote();
+        DataUtils.filterDeleted(notes);
+        return notes;
+    }
+    
+    public DishNote addDishNote(DishNote note) {
+        DishNote old = menuService.getDishNoteByName(note.getName());
+        if (old!= null && !old.isDeleted()) {
+            throw new DataConflictException("该菜品备注已存在");
+        }
+        long now = System.currentTimeMillis();
+        note.setUpdateTime(now);
+        if (old != null) {
+            note.setId(old.getId());
+            note.setCreatedTime(old.getCreatedTime());
+            menuService.updateDishNote(note);
+        } else {
+            note.setUpdateTime(now);
+            menuService.addDishNote(note);
+        }
+        return menuService.getDishNote(note.getId());
+    }
+    public DishNote updateDishNote(DishNote note) {
+        DishNote old = menuService.getDishNoteByName(note.getName());
+        if (old!= null && old.getId()!=note.getId() && !old.isDeleted()) {
+            throw new DataConflictException("该菜品备注已存在");
+        }
+        note.setUpdateTime(System.currentTimeMillis());
+        menuService.updateDishNote(note);
+        return menuService.getDishNote(note.getId());
+    }
+    public void deleteDishNote(int id) {
+        menuService.deleteDishNote(id);
     }
 }
